Consolidate GKE app create button tests
