+++ ui/src/app/components/cromwell-configuration-panel.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { AppStatus, DisksApi, WorkspaceAccessLevel } from 'generated/fetch';
+import { DisksApi, WorkspaceAccessLevel } from 'generated/fetch';
 import { AppsApi } from 'generated/fetch/api';
 
 import {
@@ -15,10 +15,7 @@ import {
   mountWithRouter,
   waitOneTickAndUpdate,
 } from 'testing/react-test-helpers';
-import {
-  AppsApiStub,
-  createListAppsCromwellResponse,
-} from 'testing/stubs/apps-api-stub';
+import { AppsApiStub } from 'testing/stubs/apps-api-stub';
 import { CdrVersionsStubVariables } from 'testing/stubs/cdr-versions-api-stub';
 import { DisksApiStub } from 'testing/stubs/disks-api-stub';
 import { ProfileStubVariables } from 'testing/stubs/profile-api-stub';
@@ -26,7 +23,6 @@ import {
   workspaceStubs,
   WorkspaceStubVariables,
 } from 'testing/stubs/workspaces';
-import { ALL_GKE_APP_STATUSES, minus } from 'testing/utils';
 
 import { defaultCromwellConfig } from './apps-panel/utils';
 
@@ -106,44 +102,6 @@ describe('CromwellConfigurationPanel', () => {
       defaultCromwellConfig
     );
     expect(onClose).toHaveBeenCalledTimes(1);
-  });
-
-  const createEnabledStatuses = [AppStatus.DELETED, null, undefined];
-  const createDisabledStatuses = minus(
-    ALL_GKE_APP_STATUSES,
-    createEnabledStatuses
-  );
-
-  describe('should allow creating a Cromwell app for certain app statuses', () => {
-    test.each(createEnabledStatuses)('Status %s', async (appStatus) => {
-      const wrapper = await component({
-        gkeAppsInWorkspace: [
-          createListAppsCromwellResponse({ status: appStatus }),
-        ],
-      });
-      expect(
-        wrapper
-          .find('#Cromwell-cloud-environment-create-button')
-          .first()
-          .prop('disabled')
-      ).toBeFalsy();
-    });
-  });
-
-  describe('should allow creating a Cromwell app for certain app statuses', () => {
-    test.each(createDisabledStatuses)('Status %s', async (appStatus) => {
-      const wrapper = await component({
-        gkeAppsInWorkspace: [
-          createListAppsCromwellResponse({ status: appStatus }),
-        ],
-      });
-      expect(
-        wrapper
-          .find('#Cromwell-cloud-environment-create-button')
-          .first()
-          .prop('disabled')
-      ).toBeTruthy();
-    });
   });
 
   it('should display a cost of $0.40 per hour when running and $0.20 per hour when paused', async () => {
